api/requirements.txt
flask==3.0.2
groq==0.4.2
pinecone-client==3.2.2
requests==2.31.0